make load-ads becomes incremental
Before you had to drop the DB and recreate it to reload ads.
This would undo crawler work on external_urls.

Now the INSERT has an ON CONFLICT modifier so adverts we've
already seen are left alone.

Also, add a -X to the psql command such that a local .psqlrc can't
request that SET and Time 6.04ms be added to STDOUT thereby corrupting
adverts.csv