Graph activity for multiple utm values
For /utm_campaign_values/:utm_campaign_value_id/between,
allow start and end params in date form. Show a multi-series
graph based on that.

Order in the query is based on count first, as charts.js will
assign colours to the first 20 in the series.

More limiting options will be required.