Fix Advert#with_utm_values
The crosstab_values subselect always returns value0, value1. That means
the WHERE clause must line up, otherwise we get zero records when we look
for a value that is not at a particular UTM index when the values are
transposed. Always sort the indices first; *hash.values order is not
defined.