Fix scraper null reference error
Scraper used to look for the last link on the page, then
looked for l.facebook.com links. It still needs to fall back
to the last link on the page when there's no l.facebook.com
link – then the existing logic will warn and set
external_url = '#no-external-url' in the case of videos.