Go via funded_by for decoding campaign
This is a mess. But it is a mess that works and it would no longer
be on fire in the run-up, so it's getting deployed.

It's a mess because the tests are set up with hosts, but there
isn't always a host, whereas there's always a funded_by.campaign.

So there's a line there, just `Decoding`, which I think is largely
useless. But I'm not retooling the tests to cope, because it's
late and I'm literally eating dinner FGS. Later, if ever.

"Is anything on fire? Is it ok if it remains on fire?"

https://twitter.com/rgarner/status/1318997250018770947